REFACTORED: removed dependancies ADDED: logging

+++ src/gllandscape/GlLandscape.h
@@ -140,7 +140,7 @@ class GlLandscape : public Landscape
     void drawTexturedTriangle1 (int x, int y);
     void drawTexturedTriangle2 (int x, int y);
   
-    void draw (Vector3 &cam, int phi, int gamma); ///< draw ALL
+    void draw (Vector3 &cam, float phi, float gamma); ///< draw ALL
     /// explosions and other light sources light up the surface, calculate dynamically by CPU
 //    void calcDynamicLight (Explosion **explo, SpaceObj **cannon, SpaceObj **missile, SpaceObj **flare);
     void calcDynamicLight (SpaceObj *object, float threshold, float maxintens, float intensfac);
